// src/Types/LineWeightType.ts
export enum LineWeightType {
	ByDIPs = -4,
	Default = -3,
	ByBlock = -2,
	ByLayer = -1,
	W0 = 0,
	W5 = 5,
	W9 = 9,
	W13 = 13,
	W15 = 15,
	W18 = 18,
	W20 = 20,
	W25 = 25,
	W30 = 30,
	W35 = 35,
	W40 = 40,
	W50 = 50,
	W53 = 53,
	W60 = 60,
	W70 = 70,
	W80 = 80,
	W90 = 90,
	W100 = 100,
	W106 = 106,
	W120 = 120,
	W140 = 140,
	W158 = 158,
	W200 = 200,
	W211 = 211,
}